requirements.txt
azureml-core==1.38.1
azureml-dataset-runtime==1.38.0
azureml-pipeline-core==1.38.0
azureml-pipeline-steps==1.38.0
